requirements.txt
torch
transformers
sentence-transformers
langchain
langchain_community
langchain-huggingface
langchain_experimental
langchainhub
streamlit
langchain_openai
python-dotenv
unstructured
langchain-google-genai
pypdf
rapidocr-onnxruntime
langchain_chroma
langchain-ollama
langgraph
langsmith
wikipedia
langchain-core
rich
langchain-tavily
duckduckgo-search

protobuf==3.20.*
